ustc-1739: refactor work item outbox partition usage
