+++ shared/admin-tools/dynamodb/archive-unwanted-outbox-records.js
@@ -1,6 +1,7 @@
 // this script is going to query and then delete records so we can perform a migration
 
 const {
+  FORMATS,
   prepareDateFromString,
 } = require('../../src/business/utilities/DateHandler');
 const { DynamoDB } = require('aws-sdk');
@@ -26,7 +27,7 @@ const archiveWorkItem = async ({ section, workItem }) => {
     pk: workItem.pk,
     sk: workItem.sk,
   };
-  const skDate = prepareDateFromString(workItem.sk.S).format('YYYY_MM');
+  const skDate = prepareDateFromString(workItem.sk.S).format(FORMATS.YYYYMM);
 
   console.log(`Key to create section-outbox|${section}|${skDate}`);
   console.log(`Key to delete ${JSON.stringify(Key)}`);
